can now click out of popup from parent element
