+++ frontend/src/components/popup.js
@@ -7,11 +7,9 @@ class Popup extends Component {
 
     render() {
         return (
-            <div className="Popup">
                 <div className="PopupCard">
                     popup
                 </div>
-            </div>
         )
     }
 }
