+++ frontend/src/styles/popup.css
@@ -13,6 +13,7 @@
     color: #fff;
 }
 
+
 .PopupCard {
     margin: auto;
     background-color: aqua;
